Fixup goproxy-compliant paths before passing them to nginx
The goproxy protocol supports case-insensitive filesystems, so it
specifies that when a package is requested by a go client, any capital
letters in the package path are converted to a '!' character followed by
the lowercase version of the letter.

https://go.dev/ref/mod#goproxy-protocol

This causes problems in Goff because Goff doesn't serve the on-disk
files itself, instead using nginx's 'X-Accel-Redirect' functionality to
request that nginx serve the file. However, Goff passes a
goproxy-compliant path to nginx, which does not match the folder
structure created by the most commonly-used downloader, the gdown gist:
https://gist.github.com/BenBurnett/1ba539ffda33607205412105fafa262c

This commit modifies goff to transform the goproxy-compliant path into
the expected filesystem path so that nginx can serve it successfully.
Without this patch, developers need to manually rename all package
directories containing capital letters to a goproxy-compliant format.